test(fix test): fix test for usetransport
