+++ packages/client-graphql/src/index.ts
@@ -2,12 +2,7 @@ export { createCourierClient } from "./client";
 export { default as Banner } from "./banner";
 export { default as Brands } from "./brands";
 export { default as Events } from "./events";
-export {
-  default as Inbox,
-  IInboxMessagePreview,
-  IInboxMessage,
-  IGetInboxMessagesParams,
-} from "./inbox";
+export { default as Inbox, IGetInboxMessagesParams } from "./inbox";
 export {
   default as Preferences,
   UpdateRecipientPreferencesPayload,
