+++ packages/core/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../tsconfig-base.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./dist",
+    "declarationDir": "./typings",
+    "paths": {
+      "~/*": ["./src/*"]
+    }
+  },
+  "include": ["../types/@types/*/index.d.ts", "./src/**/*.ts"]
+}
