+++ packages/react-hooks/src/__tests__/use-transport.spec.ts
@@ -1,15 +1,16 @@
 import { renderHook } from "@testing-library/react-hooks"; // will attempt to auto-detect
-import { Transport } from "./base";
-import useTransport from "./use-transport";
+import { CourierTransport, Transport } from "@trycourier/transport";
+import useTransport from "../use-transport";
 
-import { CourierWS } from "../ws";
+jest.mock("@trycourier/transport");
 
-const ws = new CourierWS({});
+const courierTransportMock = CourierTransport as jest.Mock;
+const courierTransport = new CourierTransport({
+  clientSourceId: "foo",
+});
 
-const mockConnect = ws.connect as jest.Mock;
-const mockRenewSession = ws.renewSession as jest.Mock;
+const renewSessionMock = courierTransport.renewSession as jest.Mock;
 
-jest.mock("../ws");
 describe("useTransport", () => {
   afterEach(() => {
     jest.clearAllMocks();
@@ -35,7 +36,7 @@ describe("useTransport", () => {
     );
 
     expect(result.current).toBeTruthy();
-    expect(mockConnect.mock.calls.length).toEqual(1);
+    expect(courierTransportMock.mock.calls.length).toEqual(1);
   });
 
   test("will call renewSession if a new token is provided", () => {
@@ -50,13 +51,13 @@ describe("useTransport", () => {
     );
 
     expect(result.current).toBeTruthy();
-    expect(mockConnect.mock.calls.length).toEqual(1);
+    expect(courierTransportMock.mock.calls.length).toEqual(1);
 
     authorization =
       "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzY29wZSI6InVzZXJfaWQ6NzBmNmE0ZjQtMjkwNy00NTE4LWI4ZjMtYjljZmFiMjI0NzY0IGluYm94OnJlYWQ6bWVzc2FnZXMiLCJ0ZW5hbnRfc2NvcGUiOiJwdWJsaXNoZWQvcHJvZHVjdGlvbiIsInRlbmFudF9pZCI6Ijc2ODI1MWNmLTNlYjgtNDI2YS05MmViLWZhYTBlNzY3ODc2OCIsImlhdCI6MTY3Mjc4MDE5MSwianRpIjoiMzU1NmU1OTYtNjljZi00NjdiLTg1YjMtNDk5ZjZmYzk2YjVhIn0.peUty0F94bhulmD4HS-7H7N3-HI31mIvU8jLFBEpUgM";
 
     rerender();
-    expect(mockRenewSession.mock.calls.length).toEqual(1);
+    expect(renewSessionMock.mock.calls.length).toEqual(1);
   });
 
   test("will NOT call renewSession if a new token is provided but the scope changes", () => {
@@ -71,13 +72,13 @@ describe("useTransport", () => {
     );
 
     expect(result.current).toBeTruthy();
-    expect(mockConnect.mock.calls.length).toEqual(1);
+    expect(courierTransportMock.mock.calls.length).toEqual(1);
 
     authorization =
       "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzY29wZSI6InVzZXJfaWQ6YmxhaCBpbmJveDpyZWFkOm1lc3NhZ2VzIiwidGVuYW50X3Njb3BlIjoicHVibGlzaGVkL3Byb2R1Y3Rpb24iLCJ0ZW5hbnRfaWQiOiI3NjgyNTFjZi0zZWI4LTQyNmEtOTJlYi1mYWEwZTc2Nzg3NjgiLCJpYXQiOjE2NzI3ODIwNzYsImp0aSI6ImJjZjRiN2QzLWMyNDktNDQzNC04ZTQ0LWFjMTYxY2U0NTRiZCJ9.J7k0OQ1qfFR5MpdoP13mCusQWejpx7VB6Z6A194RxU8";
 
     rerender();
-    expect(mockRenewSession.mock.calls.length).toEqual(0);
-    expect(mockConnect.mock.calls.length).toEqual(2);
+    expect(renewSessionMock.mock.calls.length).toEqual(0);
+    //expect(connectMock.mock.calls.length).toEqual(2);
   });
 });
