+++ packages/react-hooks/src/inbox/index.ts
@@ -1,5 +1,3 @@
-import { IInboxMessagePreview } from "@trycourier/react-provider";
-
 import { IInbox } from "./types";
 export { default as useInbox } from "./use-inbox";
-export { IInbox, IInboxMessagePreview };
+export { IInbox };
