+++ packages/react-hooks/src/index.ts
@@ -4,3 +4,4 @@ export {
   UsePreferences,
 } from "./preferences/use-preferences";
 export { PreferenceSection } from "./preferences/types";
+export { default as useTransport } from "./use-transport";
