+++ packages/react-provider/package.json
@@ -15,11 +15,12 @@
   },
   "license": "ISC",
   "dependencies": {
-    "jwt-decode": "^3.1.2",
-    "@trycourier/courier-js": "^1.4.2",
     "@trycourier/client-graphql": "^6.0.0",
+    "@trycourier/core": "^6.0.0",
+    "@trycourier/courier-js": "^1.4.2",
     "@trycourier/transport": "^6.0.0",
     "buffer": "^6.0.3",
+    "jwt-decode": "^3.1.2",
     "react-use": "^17.2.1",
     "rimraf": "^3.0.2",
     "urql": "^2.0.1",
