+++ packages/transport/src/__mocks__/index.ts
@@ -1,0 +1,12 @@
+const connectMock = jest.fn();
+const renewSessionMock = jest.fn();
+
+export const CourierTransport = jest.fn().mockImplementation(() => {
+  return {
+    connect: connectMock,
+    renewSession: renewSessionMock,
+  };
+});
+
+const transportMock = jest.fn();
+export const Transport = transportMock;
