+++ packages/transport/src/index.ts
@@ -4,6 +4,5 @@ export * from "./courier";
 export * from "./base";
 export * from "./types";
 export * from "./ws";
-export { default as useTransport } from "./use-transport";
 
 export type { ITransportOptions };
